requirements.txt
asgiref==3.6.0
attrs==22.1.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.3.0
psycopg2==2.9.5
pyrsistent==0.19.2
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
